+++ lib/src/metric/metric.dart
@@ -1,5 +1,5 @@
 import 'package:ml_linalg/vector.dart';
 
 abstract class Metric {
-  double getError(MLVector predictedLabels, MLVector origLabels);
+  double getScore(MLVector predictedLabels, MLVector origLabels);
 }